Fix relocation issues
